secure auth

+++ src/app/api/categories/route.js
@@ -1,18 +1,27 @@
 import { Category } from "@/models/Category";
 import mongoose from "mongoose";
+import { isAdmin } from "../auth/[...nextauth]/route";
 
 export async function POST(req) {
   mongoose.connect(process.env.MONGO_URL);
   const { name } = await req.json();
-  const categoryDoc = await Category.create({ name });
-  return Response.json(categoryDoc);
+  if (await isAdmin()) {
+    const categoryDoc = await Category.create({ name });
+    return Response.json(categoryDoc);
+  } else {
+    return Response.json({});
+  }
 }
 
 export async function PUT(req) {
   mongoose.connect(process.env.MONGO_URL);
-  const { name, _id } = await req.json();
-  const categoryDoc = await Category.findByIdAndUpdate(_id, { name });
-  return Response.json(categoryDoc);
+  const { _id, name } = await req.json();
+  if (await isAdmin()) {
+    const categoryDoc = await Category.findByIdAndUpdate(_id, { name });
+    return Response.json(categoryDoc);
+    // await Category.updateOne({ _id }, { name });
+  }
+  return Response.json(true);
 }
 
 export async function GET() {
@@ -25,6 +34,8 @@ export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
   const _id = url.searchParams.get("_id");
-  await Category.deleteOne({ _id });
+  if (await isAdmin()) {
+    await Category.deleteOne({ _id });
+  }
   return Response.json(true);
 }
